--- src/main/kotlin/com/chatandpay/api/dto/DepositWalletRequestDTO.kt
@@ -1,11 +1,0 @@
-package com.chatandpay.api.dto
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class DepositWalletRequestDTO (
-    val depositMoney: Int,
-    val accountId: Long,
-)
