--- src/main/kotlin/com/chatandpay/api/dto/DepositWalletResponseDTO.kt
@@ -1,11 +1,0 @@
-package com.chatandpay.api.dto
-
-import com.chatandpay.api.common.ApiResponse
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class DepositWalletResponseDTO (
-    val chargeAmount : Int,
-    val walletAmount : Int
-) : ApiResponse()
