--- src/main/kotlin/com/chatandpay/api/dto/OtherBankAccountRequestDTO.kt
@@ -1,13 +1,0 @@
-package com.chatandpay.api.dto
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
-data class OtherBankAccountRequestDTO(
-    val bankCode: String,
-    val accountNumber: String,
-    val accountName: String,
-    val autoDebitAgree: String,
-    val userId: Long,
-)
